requirements.txt
# Legal Intelligence AI System - Dependencies
# Only packages actually used in the codebase

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Data Validation
pydantic>=2.10.0

# Environment Configuration
python-dotenv>=1.0.0

# Google Gen AI SDK (Modern, replaces deprecated vertexai.generative_models)
google-genai>=1.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
